Backend/basic rest controllers

+++ backend/src/main/java/pl/juhas/backend/reservation/ReservationRepository.java
@@ -2,6 +2,6 @@ package pl.juhas.backend.reservation;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReservationRepository extends JpaRepository<Reservation, Integer> {
+public interface ReservationRepository extends JpaRepository<Reservation, Long> {
 
 }
